Set Content-Type metadata when uploading to S3
This will ensure that the objects in the bucket have a Content-Type
of "image/jpeg" in the metadata. By default, they seemed to have a
Content-Type of "application/octet-stream", and iMessage was not
displaying images in the link previews. Ensuring the image MIME type
makes the images appear in the previews.

Closes #13